--- a/bin/uglifyjs
+++ b/bin/uglifyjs
@@ -10,6 +10,7 @@ var info = require("../package.json");
 var path = require("path");
 var program = require("commander");
 var UglifyJS = require("../tools/node");
+var AST = UglifyJS.AST;
 
 var skip_keys = [ "cname", "inlined", "parent_scope", "scope", "uses_eval", "uses_with" ];
 var files = {};
@@ -182,11 +183,11 @@ if (program.self) {
 }
 
 function convert_ast(fn) {
-    return UglifyJS.ast.AST_Node.from_mozilla_ast(Object.keys(files).reduce(fn, null));
+    return AST.Node.from_mozilla_ast(Object.keys(files).reduce(fn, null));
 }
 
 function run() {
-    UglifyJS.ast.AST_Node.warn_function = function(msg) {
+    AST.Node.warn_function = function(msg) {
         print_error("WARN: " + msg);
     };
     var content = program.sourceMap && program.sourceMap.content;
@@ -260,11 +261,11 @@ function run() {
                 return value.size() ? value.map(symdef) : undefined;
             }
             if (skip_key(key)) return;
-            if (value instanceof UglifyJS.ast.AST_Token) return;
+            if (value instanceof AST.Token) return;
             if (value instanceof UglifyJS.Dictionary) return;
-            if (value instanceof UglifyJS.ast.AST_Node) {
+            if (value instanceof AST.Node) {
                 var result = {
-                    _class: "AST_" + value.TYPE
+                    _class: "AST." + value.TYPE
                 };
                 value.CTOR.PROPS.forEach(function(prop) {
                     result[prop] = value[prop];
@@ -359,27 +360,27 @@ function parse_js(flag) {
                     code: false
                 }
             }).ast.walk(new UglifyJS.TreeWalker(function(node) {
-                if (node instanceof UglifyJS.ast.AST_Assign) {
+                if (node instanceof AST.Assign) {
                     var name = node.left.print_to_string();
                     var value = node.right;
                     if (flag) {
                         options[name] = value;
-                    } else if (value instanceof UglifyJS.ast.AST_Array) {
+                    } else if (value instanceof AST.Array) {
                         options[name] = value.elements.map(to_string);
                     } else {
                         options[name] = to_string(value);
                     }
                     return true;
                 }
-                if (node instanceof UglifyJS.ast.AST_Symbol || node instanceof UglifyJS.ast.AST_PropAccess) {
+                if (node instanceof AST.Symbol || node instanceof AST.PropAccess) {
                     var name = node.print_to_string();
                     options[name] = true;
                     return true;
                 }
-                if (!(node instanceof UglifyJS.ast.AST_Sequence)) throw node;
+                if (!(node instanceof AST.Sequence)) throw node;
 
                 function to_string(value) {
-                    return value instanceof UglifyJS.ast.AST_Constant ? value.getValue() : value.print_to_string({
+                    return value instanceof AST.Constant ? value.getValue() : value.print_to_string({
                         quote_keys: true
                     });
                 }