--- a/lib/ast.js
+++ b/lib/ast.js
@@ -43,6 +43,13 @@
 
 "use strict";
 
+var utils = require("./utils");
+var noop = utils.noop;
+var string_template = utils.string_template;
+var HOP = utils.HOP;
+var MAP = utils.MAP;
+var merge = utils.merge;
+
 function DEFNODE(type, props, methods, base) {
     if (typeof base === "undefined") base = AST_Node;
     props = props ? props.split(/\s+/) : [];
@@ -323,34 +330,6 @@ var AST_Toplevel = DEFNODE("Toplevel", "globals", {
     $documentation: "The toplevel scope",
     $propdoc: {
         globals: "[Object/S] a map of name -> SymbolDef for all undeclared names",
-    },
-    wrap_commonjs: function(name) {
-        var body = this.body;
-        var wrapped_tl = "(function(exports){'$ORIG';})(typeof " + name + "=='undefined'?(" + name + "={}):" + name + ");";
-        wrapped_tl = parse(wrapped_tl);
-        wrapped_tl = wrapped_tl.transform(new TreeTransformer(function before(node) {
-            if (node instanceof AST_Directive && node.value == "$ORIG") {
-                return MAP.splice(body);
-            }
-        }));
-        return wrapped_tl;
-    },
-    wrap_enclose: function(args_values) {
-        if (typeof args_values != "string") args_values = "";
-        var index = args_values.indexOf(":");
-        if (index < 0) index = args_values.length;
-        var body = this.body;
-        return parse([
-            "(function(",
-            args_values.slice(0, index),
-            '){"$ORIG"})(',
-            args_values.slice(index + 1),
-            ")"
-        ].join("")).transform(new TreeTransformer(function(node) {
-            if (node instanceof AST_Directive && node.value == "$ORIG") {
-                return MAP.splice(body);
-            }
-        }));
     }
 }, AST_Scope);
 
@@ -952,3 +931,266 @@ TreeWalker.prototype = {
         }
     }
 };
+
+function TreeTransformer(before, after) {
+    TreeWalker.call(this);
+    this.before = before;
+    this.after = after;
+}
+TreeTransformer.prototype = new TreeWalker;
+
+(function(DEF) {
+    function do_list(list, tw) {
+        return MAP(list, function(node) {
+            return node.transform(tw, true);
+        });
+    }
+
+    DEF(AST_Node, noop);
+    DEF(AST_LabeledStatement, function(self, tw) {
+        self.label = self.label.transform(tw);
+        self.body = self.body.transform(tw);
+    });
+    DEF(AST_SimpleStatement, function(self, tw) {
+        self.body = self.body.transform(tw);
+    });
+    DEF(AST_Block, function(self, tw) {
+        self.body = do_list(self.body, tw);
+    });
+    DEF(AST_Do, function(self, tw) {
+        self.body = self.body.transform(tw);
+        self.condition = self.condition.transform(tw);
+    });
+    DEF(AST_While, function(self, tw) {
+        self.condition = self.condition.transform(tw);
+        self.body = self.body.transform(tw);
+    });
+    DEF(AST_For, function(self, tw) {
+        if (self.init) self.init = self.init.transform(tw);
+        if (self.condition) self.condition = self.condition.transform(tw);
+        if (self.step) self.step = self.step.transform(tw);
+        self.body = self.body.transform(tw);
+    });
+    DEF(AST_ForIn, function(self, tw) {
+        self.init = self.init.transform(tw);
+        self.object = self.object.transform(tw);
+        self.body = self.body.transform(tw);
+    });
+    DEF(AST_With, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+        self.body = self.body.transform(tw);
+    });
+    DEF(AST_Exit, function(self, tw) {
+        if (self.value) self.value = self.value.transform(tw);
+    });
+    DEF(AST_LoopControl, function(self, tw) {
+        if (self.label) self.label = self.label.transform(tw);
+    });
+    DEF(AST_If, function(self, tw) {
+        self.condition = self.condition.transform(tw);
+        self.body = self.body.transform(tw);
+        if (self.alternative) self.alternative = self.alternative.transform(tw);
+    });
+    DEF(AST_Switch, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+        self.body = do_list(self.body, tw);
+    });
+    DEF(AST_Case, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+        self.body = do_list(self.body, tw);
+    });
+    DEF(AST_Try, function(self, tw) {
+        self.body = do_list(self.body, tw);
+        if (self.bcatch) self.bcatch = self.bcatch.transform(tw);
+        if (self.bfinally) self.bfinally = self.bfinally.transform(tw);
+    });
+    DEF(AST_Catch, function(self, tw) {
+        self.argname = self.argname.transform(tw);
+        self.body = do_list(self.body, tw);
+    });
+    DEF(AST_Definitions, function(self, tw) {
+        self.definitions = do_list(self.definitions, tw);
+    });
+    DEF(AST_VarDef, function(self, tw) {
+        self.name = self.name.transform(tw);
+        if (self.value) self.value = self.value.transform(tw);
+    });
+    DEF(AST_Lambda, function(self, tw) {
+        if (self.name) self.name = self.name.transform(tw);
+        self.argnames = do_list(self.argnames, tw);
+        self.body = do_list(self.body, tw);
+    });
+    DEF(AST_Call, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+        self.args = do_list(self.args, tw);
+    });
+    DEF(AST_Sequence, function(self, tw) {
+        self.expressions = do_list(self.expressions, tw);
+    });
+    DEF(AST_Dot, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+    });
+    DEF(AST_Sub, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+        self.property = self.property.transform(tw);
+    });
+    DEF(AST_Unary, function(self, tw) {
+        self.expression = self.expression.transform(tw);
+    });
+    DEF(AST_Binary, function(self, tw) {
+        self.left = self.left.transform(tw);
+        self.right = self.right.transform(tw);
+    });
+    DEF(AST_Conditional, function(self, tw) {
+        self.condition = self.condition.transform(tw);
+        self.consequent = self.consequent.transform(tw);
+        self.alternative = self.alternative.transform(tw);
+    });
+    DEF(AST_Array, function(self, tw) {
+        self.elements = do_list(self.elements, tw);
+    });
+    DEF(AST_Object, function(self, tw) {
+        self.properties = do_list(self.properties, tw);
+    });
+    DEF(AST_ObjectProperty, function(self, tw) {
+        self.value = self.value.transform(tw);
+    });
+})(function(node, descend) {
+    node.DEFMETHOD("transform", function(tw, in_list) {
+        var x, y;
+        tw.push(this);
+        if (tw.before) x = tw.before(this, descend, in_list);
+        if (typeof x === "undefined") {
+            x = this;
+            descend(x, tw);
+            if (tw.after) {
+                y = tw.after(x, in_list);
+                if (typeof y !== "undefined") x = y;
+            }
+        }
+        tw.pop();
+        return x;
+    });
+});
+
+/***********************************************************************
+/* Helpers
+/**********************************************************************/
+
+// return true if the node at the top of the stack (that means the
+// innermost node in the current output) is lexically the first in
+// a statement.
+function first_in_statement(stack) {
+    var node = stack.parent(-1);
+    for (var i = 0, p; p = stack.parent(i++); node = p) {
+        if (p.TYPE == "Call") {
+            if (p.expression === node) continue;
+        } else if (p instanceof AST_Binary) {
+            if (p.left === node) continue;
+        } else if (p instanceof AST_Conditional) {
+            if (p.condition === node) continue;
+        } else if (p instanceof AST_PropAccess) {
+            if (p.expression === node) continue;
+        } else if (p instanceof AST_Sequence) {
+            if (p.expressions[0] === node) continue;
+        } else if (p instanceof AST_Statement) {
+            return p.body === node;
+        } else if (p instanceof AST_UnaryPostfix) {
+            if (p.expression === node) continue;
+        }
+        return false;
+    }
+}
+
+merge(exports, {
+    DEFNODE : DEFNODE,
+    AST_Token : AST_Token,
+    AST_Node : AST_Node,
+    AST_Statement : AST_Statement,
+    AST_Debugger : AST_Debugger,
+    AST_Directive : AST_Directive,
+    AST_SimpleStatement : AST_SimpleStatement,
+    walk_body : walk_body,
+    AST_Block : AST_Block,
+    AST_BlockStatement : AST_BlockStatement,
+    AST_EmptyStatement : AST_EmptyStatement,
+    AST_StatementWithBody : AST_StatementWithBody,
+    AST_LabeledStatement : AST_LabeledStatement,
+    AST_IterationStatement : AST_IterationStatement,
+    AST_DWLoop : AST_DWLoop,
+    AST_Do : AST_Do,
+    AST_While : AST_While,
+    AST_For : AST_For,
+    AST_ForIn : AST_ForIn,
+    AST_With : AST_With,
+    AST_Scope : AST_Scope,
+    AST_Toplevel : AST_Toplevel,
+    AST_Lambda : AST_Lambda,
+    AST_Accessor : AST_Accessor,
+    AST_Function : AST_Function,
+    AST_Defun : AST_Defun,
+    AST_Jump : AST_Jump,
+    AST_Exit : AST_Exit,
+    AST_Return : AST_Return,
+    AST_Throw : AST_Throw,
+    AST_LoopControl : AST_LoopControl,
+    AST_Break : AST_Break,
+    AST_Continue : AST_Continue,
+    AST_If : AST_If,
+    AST_Switch : AST_Switch,
+    AST_SwitchBranch : AST_SwitchBranch,
+    AST_Default : AST_Default,
+    AST_Case : AST_Case,
+    AST_Try : AST_Try,
+    AST_Catch : AST_Catch,
+    AST_Finally : AST_Finally,
+    AST_Definitions : AST_Definitions,
+    AST_Var : AST_Var,
+    AST_VarDef : AST_VarDef,
+    AST_Call : AST_Call,
+    AST_New : AST_New,
+    AST_Sequence : AST_Sequence,
+    AST_PropAccess : AST_PropAccess,
+    AST_Dot : AST_Dot,
+    AST_Sub : AST_Sub,
+    AST_Unary : AST_Unary,
+    AST_UnaryPrefix : AST_UnaryPrefix,
+    AST_UnaryPostfix : AST_UnaryPostfix,
+    AST_Binary : AST_Binary,
+    AST_Conditional : AST_Conditional,
+    AST_Assign : AST_Assign,
+    AST_Array : AST_Array,
+    AST_Object : AST_Object,
+    AST_ObjectProperty : AST_ObjectProperty,
+    AST_ObjectKeyVal : AST_ObjectKeyVal,
+    AST_ObjectSetter : AST_ObjectSetter,
+    AST_ObjectGetter : AST_ObjectGetter,
+    AST_Symbol : AST_Symbol,
+    AST_SymbolAccessor : AST_SymbolAccessor,
+    AST_SymbolDeclaration : AST_SymbolDeclaration,
+    AST_SymbolVar : AST_SymbolVar,
+    AST_SymbolFunarg : AST_SymbolFunarg,
+    AST_SymbolDefun : AST_SymbolDefun,
+    AST_SymbolLambda : AST_SymbolLambda,
+    AST_SymbolCatch : AST_SymbolCatch,
+    AST_Label : AST_Label,
+    AST_SymbolRef : AST_SymbolRef,
+    AST_LabelRef : AST_LabelRef,
+    AST_This : AST_This,
+    AST_Constant : AST_Constant,
+    AST_String : AST_String,
+    AST_Number : AST_Number,
+    AST_RegExp : AST_RegExp,
+    AST_Atom : AST_Atom,
+    AST_Null : AST_Null,
+    AST_NaN : AST_NaN,
+    AST_Undefined : AST_Undefined,
+    AST_Hole : AST_Hole,
+    AST_Infinity : AST_Infinity,
+    AST_Boolean : AST_Boolean,
+    AST_False : AST_False,
+    AST_True : AST_True,
+    TreeWalker : TreeWalker,
+    TreeTransformer : TreeTransformer,
+    first_in_statement : first_in_statement,
+});