--- a/lib/compress.js
+++ b/lib/compress.js
@@ -43,6 +43,115 @@
 
 "use strict";
 
+var utils = require("./utils");
+var member = utils.member;
+var defaults = utils.defaults;
+var merge = utils.merge;
+var noop = utils.noop;
+var return_false = utils.return_false;
+var return_true = utils.return_true;
+var return_this = utils.return_this;
+var return_null = utils.return_null;
+var string_template = utils.string_template;
+var remove = utils.remove;
+var makePredicate = utils.makePredicate;
+var all = utils.all;
+var Dictionary = utils.Dictionary;
+var HOP = utils.HOP;
+var MAP = utils.MAP;
+
+var ast = require("./ast");
+var walk_body = ast.walk_body;
+var TreeWalker = ast.TreeWalker;
+var TreeTransformer = ast.TreeTransformer;
+var first_in_statement = ast.first_in_statement;
+var AST_Node = ast.AST_Node;
+var AST_Statement = ast.AST_Statement;
+var AST_Debugger = ast.AST_Debugger;
+var AST_Directive = ast.AST_Directive;
+var AST_SimpleStatement = ast.AST_SimpleStatement;
+var AST_Block = ast.AST_Block;
+var AST_BlockStatement = ast.AST_BlockStatement;
+var AST_EmptyStatement = ast.AST_EmptyStatement;
+var AST_LabeledStatement = ast.AST_LabeledStatement;
+var AST_IterationStatement = ast.AST_IterationStatement;
+var AST_DWLoop = ast.AST_DWLoop;
+var AST_Do = ast.AST_Do;
+var AST_While = ast.AST_While;
+var AST_For = ast.AST_For;
+var AST_ForIn = ast.AST_ForIn;
+var AST_With = ast.AST_With;
+var AST_Scope = ast.AST_Scope;
+var AST_Toplevel = ast.AST_Toplevel;
+var AST_Lambda = ast.AST_Lambda;
+var AST_Accessor = ast.AST_Accessor;
+var AST_Function = ast.AST_Function;
+var AST_Defun = ast.AST_Defun;
+var AST_Jump = ast.AST_Jump;
+var AST_Exit = ast.AST_Exit;
+var AST_Return = ast.AST_Return;
+var AST_LoopControl = ast.AST_LoopControl;
+var AST_Break = ast.AST_Break;
+var AST_Continue = ast.AST_Continue;
+var AST_If = ast.AST_If;
+var AST_Switch = ast.AST_Switch;
+var AST_SwitchBranch = ast.AST_SwitchBranch;
+var AST_Default = ast.AST_Default;
+var AST_Case = ast.AST_Case;
+var AST_Try = ast.AST_Try;
+var AST_Catch = ast.AST_Catch;
+var AST_Finally = ast.AST_Finally;
+var AST_Definitions = ast.AST_Definitions;
+var AST_Var = ast.AST_Var;
+var AST_VarDef = ast.AST_VarDef;
+var AST_Call = ast.AST_Call;
+var AST_New = ast.AST_New;
+var AST_Sequence = ast.AST_Sequence;
+var AST_PropAccess = ast.AST_PropAccess;
+var AST_Dot = ast.AST_Dot;
+var AST_Sub = ast.AST_Sub;
+var AST_Unary = ast.AST_Unary;
+var AST_UnaryPrefix = ast.AST_UnaryPrefix;
+var AST_UnaryPostfix = ast.AST_UnaryPostfix;
+var AST_Binary = ast.AST_Binary;
+var AST_Conditional = ast.AST_Conditional;
+var AST_Assign = ast.AST_Assign;
+var AST_Array = ast.AST_Array;
+var AST_Object = ast.AST_Object;
+var AST_ObjectProperty = ast.AST_ObjectProperty;
+var AST_ObjectKeyVal = ast.AST_ObjectKeyVal;
+var AST_Symbol = ast.AST_Symbol;
+var AST_SymbolDeclaration = ast.AST_SymbolDeclaration;
+var AST_SymbolVar = ast.AST_SymbolVar;
+var AST_SymbolFunarg = ast.AST_SymbolFunarg;
+var AST_SymbolDefun = ast.AST_SymbolDefun;
+var AST_SymbolLambda = ast.AST_SymbolLambda;
+var AST_SymbolCatch = ast.AST_SymbolCatch;
+var AST_SymbolRef = ast.AST_SymbolRef;
+var AST_This = ast.AST_This;
+var AST_Constant = ast.AST_Constant;
+var AST_String = ast.AST_String;
+var AST_Number = ast.AST_Number;
+var AST_RegExp = ast.AST_RegExp;
+var AST_Null = ast.AST_Null;
+var AST_NaN = ast.AST_NaN;
+var AST_Undefined = ast.AST_Undefined;
+var AST_Hole = ast.AST_Hole;
+var AST_Infinity = ast.AST_Infinity;
+var AST_Boolean = ast.AST_Boolean;
+var AST_False = ast.AST_False;
+var AST_True = ast.AST_True;
+
+var parse = require("./parse");
+var is_identifier_string = parse.is_identifier_string;
+var JS_Parse_Error = parse.JS_Parse_Error;
+var PRECEDENCE = parse.PRECEDENCE;
+var parse = parse.parse;
+
+var OutputStream = require("./output").OutputStream;
+
+require("./scope")
+
 function Compressor(options, false_by_default) {
     if (!(this instanceof Compressor))
         return new Compressor(options, false_by_default);
@@ -6400,4 +6509,8 @@ merge(Compressor.prototype, {
         opt._optimized = true;
         return opt;
     });
+});
+
+merge(exports, {
+    Compressor : Compressor,
 });