--- a/lib/minify.js
+++ b/lib/minify.js
@@ -1,5 +1,30 @@
 "use strict";
 
+var utils = require("./utils");
+var defaults = utils.defaults;
+var Dictionary = utils.Dictionary;
+var HOP = utils.HOP;
+var MAP = utils.MAP;
+var merge = utils.merge;
+
+var ast = require("./ast");
+var AST_Node = ast.AST_Node;
+var AST_Toplevel = ast.AST_Toplevel;
+var AST_Directive = ast.AST_Directive;
+var TreeTransformer = ast.TreeTransformer;
+
+var parse = require("./parse");
+var parse = parse.parse;
+
+var OutputStream = require("./output").OutputStream;
+var Compressor = require("./compress").Compressor;
+
+var SourceMap = require("./sourcemap").SourceMap;
+
+var propmangle = require("./propmangle");
+var reserve_quoted_keys = propmangle.reserve_quoted_keys;
+var mangle_properties = propmangle.mangle_properties;
+
 var to_ascii = typeof atob == "undefined" ? function(b64) {
     return new Buffer(b64, "base64").toString();
 } : atob;
@@ -51,6 +76,35 @@ function to_json(cache) {
 }
 
 function minify(files, options) {
+    function wrap_commonjs(toplevel_node, name) {
+        var body = toplevel_node.body;
+        var wrapped_tl = "(function(exports){'$ORIG';})(typeof " + name + "=='undefined'?(" + name + "={}):" + name + ");";
+        wrapped_tl = parse(wrapped_tl);
+        wrapped_tl = wrapped_tl.transform(new TreeTransformer(function before(node) {
+            if (node instanceof AST_Directive && node.value == "$ORIG") {
+                return MAP.splice(body);
+            }
+        }));
+        return wrapped_tl;
+    }
+    function wrap_enclose(toplevel_node, args_values) {
+        if (typeof args_values != "string") args_values = "";
+        var index = args_values.indexOf(":");
+        if (index < 0) index = args_values.length;
+        var body = toplevel_node.body;
+        return parse([
+            "(function(",
+            args_values.slice(0, index),
+            '){"$ORIG"})(',
+            args_values.slice(index + 1),
+            ")"
+        ].join("")).transform(new TreeTransformer(function(node) {
+            if (node instanceof AST_Directive && node.value == "$ORIG") {
+                return MAP.splice(body);
+            }
+        }));
+    }
+    
     var warn_function = AST_Node.warn_function;
     try {
         options = defaults(options, {
@@ -156,10 +210,10 @@ function minify(files, options) {
             reserve_quoted_keys(toplevel, quoted_props);
         }
         if (options.wrap) {
-            toplevel = toplevel.wrap_commonjs(options.wrap);
+            toplevel = wrap_commonjs(toplevel, options.wrap);
         }
         if (options.enclose) {
-            toplevel = toplevel.wrap_enclose(options.enclose);
+            toplevel = wrap_enclose(toplevel, options.enclose);
         }
         if (timings) timings.rename = Date.now();
         if (options.rename) {
@@ -244,3 +298,14 @@ function minify(files, options) {
         AST_Node.warn_function = warn_function;
     }
 }
+
+merge(exports, {
+    to_ascii : to_ascii,
+    to_base64 : to_base64,
+    read_source_map : read_source_map,
+    parse_source_map : parse_source_map,
+    set_shorthand : set_shorthand,
+    init_cache : init_cache,
+    to_json : to_json,
+    minify : minify,
+});