--- a/lib/propmangle.js
+++ b/lib/propmangle.js
@@ -43,6 +43,27 @@
 
 "use strict";
 
+var utils = require("./utils");
+var defaults = utils.defaults;
+var push_uniq = utils.push_uniq;
+var Dictionary = utils.Dictionary;
+var merge = utils.merge;
+
+var scope = require("./scope");
+var base54 = scope.base54;
+
+var ast = require("./ast");
+var TreeWalker = ast.TreeWalker;
+var TreeTransformer = ast.TreeTransformer;
+var AST_Call = ast.AST_Call;
+var AST_Sequence = ast.AST_Sequence;
+var AST_Dot = ast.AST_Dot;
+var AST_Sub = ast.AST_Sub;
+var AST_Conditional = ast.AST_Conditional;
+var AST_ObjectProperty = ast.AST_ObjectProperty;
+var AST_ObjectKeyVal = ast.AST_ObjectKeyVal;
+var AST_String = ast.AST_String;
+
 function find_builtins(reserved) {
     // NaN will be included due to Number.NaN
     [
@@ -250,3 +271,10 @@ function mangle_properties(ast, options) {
         }));
     }
 }
+
+merge(exports, {
+    find_builtins : find_builtins,
+    reserve_quoted_keys : reserve_quoted_keys,
+    addStrings : addStrings,
+    mangle_properties : mangle_properties,
+});