--- a/lib/sourcemap.js
+++ b/lib/sourcemap.js
@@ -43,6 +43,11 @@
 
 "use strict";
 
+var MOZ_SourceMap = require("source-map");
+var utils = require("./utils");
+var defaults = utils.defaults;
+var merge = utils.merge;
+
 // a small wrapper around fitzgen's source-map library
 function SourceMap(options) {
     options = defaults(options, {
@@ -102,3 +107,7 @@ function SourceMap(options) {
         }
     };
 }
+
+merge(exports, {
+    SourceMap : SourceMap,
+});