--- a/test/mocha.js
+++ b/test/mocha.js
@@ -22,6 +22,7 @@ it = function(title, fn) {
     tasks.push(fn);
 };
 
+
 fs.readdirSync("test/mocha").filter(function(file) {
     return /\.js$/.test(file);
 }).forEach(function(file) {