--- a/tools/exports.js
+++ b/tools/exports.js
@@ -1,6 +1,8 @@
+exports["Compressor"] = Compressor;
 exports["Dictionary"] = Dictionary;
 exports["TreeWalker"] = TreeWalker;
 exports["TreeTransformer"] = TreeTransformer;
 exports["minify"] = minify;
 exports["parse"] = parse;
 exports["_push_uniq"] = push_uniq;
+exports["string_template"] = string_template;