… README.md
… lib/mozilla-ast.js
… lib/output.js
… lib/parse.js
… lib/scope.js
… lib/transform.js
… lib/utils.js
… test/exports.js
… test/mocha/cli.js
… test/mocha/spidermonkey.js
… test/node.js
… tools/generate-exports.js
… tools/node.js