This closes #3181 by using the lib-files as modules.

Details:

- I created a script to generate the exports statement from every top-level variable and function definition (careful when running it, it changes files!)
- `exports.foo` is used because  
  - shorter
  - works with VSC: https://github.com/Microsoft/vscode/issues/51520
  - `exports = {...}` would break the standalone package
- Add `var foo = <import>.foo` statements to avoid code changes
- Moved some functions (e.g. removed transform.js) to fix circular dependencies

TODO improvements:

- Rather use `<import>.foo` directly instead of defining dozens of variables except for e.g. the utils. This would also be cleaner as it is clear where a function comes from
- I would replace every use of `AST_<someNode>`/`ast.AST_<someNode>` by `AST.<someNode>` (fix exports accordingly), this saves the imports (see first point) and looks the same

Timings (single run):

Master:

> https://code.jquery.com/jquery-3.2.1.js
> - total: 5.017s
> https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.4/angular.js
> - total: 8.753s
> https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
> - total: 14.084s
> https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
> - total: 1.946s
> https://unpkg.com/react@15.3.2/dist/react.js
> - total: 7.977s
> http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
> - total: 11.541s
> https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
> - total: 7.046s
> https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
> - total: 9.889s
> https://raw.githubusercontent.com/kangax/html-minifier/v3.5.7/dist/htmlminifier.js
> - total: 10.694s

This PR:

> https://code.jquery.com/jquery-3.2.1.js
> - total: 3.467s
> https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.6.4/angular.js
> - total: 6.483s
> https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
> - total: 10.613s
> https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
> - total: 0.664s
> https://unpkg.com/react@15.3.2/dist/react.js
> - total: 6.108s
> http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
> - total: 6.866s
> https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
> - total: 5.447s
> https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
> - total: 8.550s
> https://raw.githubusercontent.com/kangax/html-minifier/v3.5.7/dist/htmlminifier.js
> - total: 11.831s

Note how it mostly got faster!