@fabiosantoscode 2018-06-11
Awesome work right here! I wasn't expecting there to be a speedup at all, so this surprised me.

I also agree that ast.NodeName is better than AST_NodeName, removes a lot of those variables.

Commenting so I get updates on this PR. Juicy stuff!