@kzc 2018-06-12
> I also agree that ast.NodeName is better than AST_NodeName, removes a lot of those variables.

For backwards compatibility with external user traverse/transform code, the module should still export the `AST_` prefixed node class names. 

Whether we use per-source-file class aliases internally depends on whether there are any performance implications for `ast.NodeClass` vs `NodeClass` in `lib/compress.js` - which can be easily tested on large JS inputs.