@Flamefire 2018-06-12
I see no problem exporting them as both and would deprecate `AST_NodeName` to remove it later. Prefixing was necessary as everything was put into 1 namespace but limiting the scope and using proper names is actually the better solution.

I could commit the change and run the benchmark again.