@kzc 2018-06-12
> deprecate AST_NodeName to remove it later. 

We wouldn't want to deprecate the `AST_` prefix for exports. It distinguishes those symbols as being AST classes, which has value.