@kzc 2018-06-13
Does this Function "eval" impede debugability?

https://github.com/Flamefire/UglifyJS2/blob/13196026ebf72a46553c9bd1a1d769b052b34aeb/lib/ast.js#L68