@Flamefire 2018-06-13
> We wouldn't want to deprecate the AST_ prefix for exports. It distinguishes those symbols as being AST classes, which has value.

What I want to propose is using `AST.Node` directly after an import `var AST=require(...ast)` or from outside: `UglifyJS.AST.Node`. See my latest commit which does exactly that. I also added code to not break current code by providing  `UglifyJS.AST_Node` too although I would advice to drop that. It might even affect performance.

Speaking of (laptop):

Master:
- total: 11.355s
- total: 20.531s
- total: 31.374s
- total: 4.096s
- total: 18.878s
- total: 29.394s
- total: 16.955s
- total: 24.268s
- total: 29.994s

Commit 1:
- total: 11.211s
- total: 19.263s
- total: 29.751s
- total: 4.176s
- total: 18.952s
- total: 27.922s
- total: 16.660s
- total: 22.638s
- total: 27.448s

Commit 2:
- total: 12.053s
- total: 20.546s
- total: 34.570s
- total: 3.857s
- total: 19.542s
- total: 31.131s
- total: 16.809s
- total: 24.939s
- total: 30.193s

Executing the benchmark 3 times and summing it up doesn't look good:

- Master: 472.396s
- Commit1: 463.934s
- Commit2: 570.292s