@kzc 2018-06-13
Are you sure about the 570.292s figure? The sum of the timings per run above are:

* master: 186.845
* commit1: 178.021
* commit2: 193.64

Granted the `AST.NodeClass` figures in commit2 are worse than the other two runs, but not 25% worse.

It appears that the local class aliases are the most efficient way to go.