@Flamefire 2018-06-13
> Does this Function "eval" impede debugability?

Yes. It is not possible to set persistent breakpoints on eval'd code as the code is removed and regenerated. However VSC **can**  step into eval'd code. So as long as the eval thing is small or the benefit of the eval is large, it is a ok-tradeoff.

Yes I'm sure. I have no idea why though. Commands used:
```
(node test/benchmark.js && node test/benchmark.js && node test/benchmark.js) > ~/timing-PR2-2.txt
(grep total: ~/timing-PR2-2.txt | awk '{print $3}'| sed 's/s/ + /' | tr -d '\n';echo 0) | bc
```

> Master: 
> - total: 9.662s
> - total: 16.734s
> - total: 26.128s
> - total: 3.656s
> - total: 16.487s
> - total: 23.311s
> - total: 14.045s
> - total: 20.027s
> - total: 25.024s
> - total: 9.428s
> - total: 17.961s
> - total: 26.900s
> - total: 3.401s
> - total: 16.986s
> - total: 23.759s
> - total: 14.611s
> - total: 20.858s
> - total: 25.762s
> - total: 10.164s
> - total: 16.947s
> - total: 26.169s
> - total: 3.653s
> - total: 16.318s
> - total: 24.654s
> - total: 14.446s
> - total: 20.084s
> - total: 25.221s
> 
> PR2-2:
> - total: 11.132s
> - total: 19.854s
> - total: 31.990s
> - total: 3.976s
> - total: 18.980s
> - total: 28.244s
> - total: 16.625s
> - total: 24.532s
> - total: 30.240s
> - total: 11.041s
> - total: 19.778s
> - total: 31.378s
> - total: 4.305s
> - total: 19.384s
> - total: 27.824s
> - total: 16.162s
> - total: 23.977s
> - total: 30.512s
> - total: 11.840s
> - total: 22.723s
> - total: 33.441s
> - total: 4.971s
> - total: 20.791s
> - total: 29.269s
> - total: 18.051s
> - total: 25.753s
> - total: 31.920s

Maybe the single master run above was "unlucky"