@kzc 2018-06-13
Either way, commit1 with its local class aliases has the best timings. `AST.ClassName` has a noticeable cost versus `AST_ClassName`.