@alexlamsl 2018-06-13
Such an extensively disruptive change is unacceptable, especially given the hacky nature of bending the code around just to use `require()`.